Initial Code Commit
This commit adds actual code. All of it is being generated through a
Yeoman generator. You can find the details on this here:
https://www.npmjs.org/package/generator-angular-jade-stylus

This will let us do all sorts of fancy stuff in the future. It sets up
an easy way to add new code and uses a tried and true application
structure and all that whatnot.

I went with CoffeeScript, Jade, and Stylus. This makes our code a lot
cleaner and easier to modify later. For transitioning from Javascript to
Coffee, check out http://js2coffee.org/